chore(ds): refactor type definitions
