+++ packages/design-systems/src/components/composite/Datagrid/Export/Export.tsx
@@ -2,8 +2,8 @@ import { useRef } from "react";
 import { mkConfig, generateCsv, download } from "export-to-csv";
 import { DownloadIcon } from "lucide-react";
 import { Table, TableFeature, makeStateUpdater } from "@tanstack/react-table";
-import { ExportOptions, ExportProps, ExportTableState } from "../types";
 import { addEventOutside } from "../addEventOutside";
+import type { ExportOptions, ExportProps, ExportTableState } from "./types";
 import { Box } from "@components/patterns/Box";
 import { HStack } from "@components/patterns/HStack";
 import { Button } from "@components/Button";
