+++ packages/design-systems/src/components/composite/Datagrid/Export/types.ts
@@ -1,0 +1,27 @@
+import { Updater } from "@tanstack/table-core";
+import { Localization } from "..";
+
+export type ExportState = {
+  enableCsvExport?: boolean;
+};
+
+export interface ExportTableState {
+  exportOptions: ExportState;
+  exportOpen: boolean;
+}
+
+export interface ExportOptions {}
+
+export interface ExportInstance {
+  setExportOpen: (updater: Updater<boolean>) => void;
+  getEnableExport: () => boolean;
+  exportCsv: () => void;
+}
+
+export type ExportProps = {
+  exportOptions?: ExportState;
+  localization: Localization;
+  exportCsv: () => void;
+  exportOpen: boolean;
+  setExportOpen: (updater: Updater<boolean>) => void;
+};
