+++ packages/design-systems/src/components/composite/Datagrid/Pagination/Pagination.tsx
@@ -11,12 +11,13 @@ import {
   ChevronDown,
   CheckIcon,
 } from "lucide-react";
-import { DataGridInstance, ValueChangeDetails } from "../types";
+import { DataGridInstance } from "../types";
 import { Text } from "../../../Text";
 import { Button } from "../../../Button";
 import { IconButton } from "../../../IconButton";
+import type { ValueChangeDetails } from "../SearchFilter/types";
 import { Select, usePagination } from "./utils";
-import { Localization } from "@locales/types";
+import type { Localization } from "@locales/types";
 
 const classes = select();
