+++ packages/design-systems/src/components/composite/Datagrid/Pagination/utils.ts
@@ -1,7 +1,12 @@
 import { useMemo } from "react";
 import { Select as AS } from "@ark-ui/react/select";
 import { styled } from "@tailor-platform/styled-system/jsx";
-import { DataGridInstance, Page } from "../types";
+import { DataGridInstance } from "../types";
+
+type Page = {
+  index: number;
+  type: "page" | "ellipsis";
+};
 
 const ELLIPSIS_SIZE = 4;
