+++ packages/design-systems/src/components/composite/Datagrid/SearchFilter/CustomFilter.test.tsx
@@ -11,12 +11,7 @@ import { ColumnDef } from "@tanstack/react-table";
 import userEvent from "@testing-library/user-event";
 import { describe, expect, it } from "vitest";
 import "@testing-library/jest-dom/vitest"; //For userEvent used for clicks etc. DONT use fireEvent for clicks etc. as it might not work properly with select elements
-import {
-  Column,
-  DataGridInstance,
-  GraphQLQueryFilter,
-  UseDataGridProps,
-} from "../types";
+import { Column, DataGridInstance, UseDataGridProps } from "../types";
 import { LOCALIZATION_JA } from "../../../../locales/ja";
 import { LOCALIZATION_EN } from "../../../../locales/en";
 import {
@@ -33,6 +28,7 @@ import { useDataGrid } from "../useDataGrid";
 import { Payment, originData, setFilterChange } from "../utils/test";
 import { DataGrid } from "../Datagrid";
 import { CustomFilter } from "./CustomFilter";
+import type { GraphQLQueryFilter } from "./types";
 
 /* eslint-disable-next-line @typescript-eslint/no-empty-function */
 window.HTMLElement.prototype.scrollTo = function () {}; //(https://github.com/jsdom/jsdom/issues/1695)
