+++ packages/design-systems/src/components/composite/Datagrid/index.tsx
@@ -1,8 +1,5 @@
 export { DataGrid } from "./Datagrid";
 export { useDataGrid } from "./useDataGrid";
-export type {
-  DataGridInstance,
-  UseDataGridProps,
-  GraphQLQueryFilter,
-} from "./types";
+export type { DataGridInstance, UseDataGridProps } from "./types";
+export type { GraphQLQueryFilter } from "./SearchFilter/types";
 export type { Localization } from "../../../locales/types";
